revert changes after the sqlite3 adapter for numpy float32

+++ hydroffice/soundspeed/atlas/woa09/woa09.py
@@ -313,7 +313,7 @@ class Woa09(AbstractAtlas):
         ssp.meta.utc_time = dt(year=datestamp.year, month=datestamp.month, day=datestamp.day)
         ssp.meta.original_path = "WOA09_%s" % datestamp.strftime("%Y%m%d")
         ssp.init_data(num_values)
-        ssp.data.depth = self.t_seasonal.variables['depth'][0:num_values].astype(np.float64)
+        ssp.data.depth = self.t_seasonal.variables['depth'][0:num_values]
         ssp.data.temp = t[valid]
         ssp.data.sal = s[valid]
         ssp.calc_data_speed()
@@ -335,7 +335,7 @@ class Woa09(AbstractAtlas):
         ssp_min.meta.utc_time = dt(year=datestamp.year, month=datestamp.month, day=datestamp.day)
         if num_values > 0:
             ssp_min.init_data(num_values)
-            ssp_min.data.depth = self.t_seasonal.variables['depth'][0:num_values].astype(np.float64)
+            ssp_min.data.depth = self.t_seasonal.variables['depth'][0:num_values]
             ssp_min.data.temp = t_min[valid][0:num_values]
             ssp_min.data.sal = s_min[valid][0:num_values]
             ssp_min.calc_data_speed()
